Add unit tests and rename tokenApi.test file
